Fix embedded OCR output when resampling.
When resampling, keep both the old resolution in $resolution (for proper
interpretation of the OCR text positions) and the new, resampled
resolution (for display).

This also slightly cleans up create_PDF by introducing a POINTS_PER_INCH
constant for extra readability, expressing $w and $h in inches instead
of pixels, and passing a width and height instead of a scale when
embedding an image.