Avoid issues when LANGUAGE is not set by using LC_MESSAGES
If the environment variable LANGUAGE is not defined:

* The user is warned they are using locale '' and that gscan2pdf does
  not currently know which tesseract language package would be necessary
  for that locale.

* --debug twice reports "Use of uninitialized value $locale".

Because only the first two characters of the first entry in LANGUAGE
are used LC_MESSAGES should be a reliable substitute. This still allows
for the case when a user has defined a different first choice message
language to their system language, if they set LC_MESSAGES.
Retrieve it using setlocale() to ensure a value is returned when the
LC_MESSAGES environment variable is not defined.