Fix parsing scanimage version
The code needs to know a scanimage version to supply --format=pnm
option requires since 1.0.27. The code used to compare only the last
number (27) and that stopped working with sane-backends-1.1.1
(1 < 27). It broke t/081_Frontend_CLI.t:

    ok 43 - scanimage starts
    not ok 44 - error callback called: Unknown message: Output format is not set, using pnm as a default.
    #   Failed test 'error callback called: Unknown message: Output format is not set, using pnm as a default.'
    #   at t/081_Frontend_CLI.t line 384.
    ok 45 - scanimage scans
    ok 46 - scanimage finishes

This patch fixes it by using version module to parse the complete
version string.

Signed-off-by: Petr Písař <ppisar@redhat.com>